Remove NextGen build cache prototype
