+++ subprojects/build-cache-http/src/main/java/org/gradle/caching/http/internal/HttpBuildCacheService.java
@@ -22,16 +22,16 @@ import org.apache.http.HttpStatus;
 import org.apache.http.StatusLine;
 import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.methods.HttpHead;
 import org.apache.http.client.methods.HttpPut;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.entity.AbstractHttpEntity;
 import org.apache.http.protocol.HTTP;
 import org.gradle.api.UncheckedIOException;
 import org.gradle.caching.BuildCacheEntryReader;
+import org.gradle.caching.BuildCacheEntryWriter;
 import org.gradle.caching.BuildCacheException;
 import org.gradle.caching.BuildCacheKey;
-import org.gradle.caching.internal.NextGenBuildCacheService;
+import org.gradle.caching.BuildCacheService;
 import org.gradle.internal.UncheckedException;
 import org.gradle.internal.resource.transport.http.HttpClientHelper;
 import org.gradle.internal.resource.transport.http.HttpClientResponse;
@@ -48,7 +48,7 @@ import java.util.Set;
 /**
  * Build cache implementation that delegates to a service accessible via HTTP.
  */
-public class HttpBuildCacheService implements NextGenBuildCacheService {
+public class HttpBuildCacheService implements BuildCacheService {
     private static final Logger LOGGER = LoggerFactory.getLogger(HttpBuildCacheService.class);
     static final String BUILD_CACHE_CONTENT_TYPE = "application/vnd.gradle.build-cache-artifact.v1";
 
@@ -73,31 +73,6 @@ public class HttpBuildCacheService implements NextGenBuildCacheService {
         this.useExpectContinue = useExpectContinue;
         this.root = withTrailingSlash(url);
         this.httpClientHelper = httpClientHelper;
-    }
-
-    @Override
-    public boolean contains(BuildCacheKey key) {
-        final URI uri = root.resolve("./" + key.getHashCode());
-        HttpHead httpHead = new HttpHead(uri);
-        requestCustomizer.customize(httpHead);
-
-        try (HttpClientResponse response = httpClientHelper.performHttpRequest(httpHead)) {
-            StatusLine statusLine = response.getStatusLine();
-            if (LOGGER.isDebugEnabled()) {
-                LOGGER.debug("Response for HEAD {}: {}", safeUri(uri), statusLine);
-            }
-            int statusCode = statusLine.getStatusCode();
-            if (isHttpSuccess(statusCode)) {
-                return true;
-            } else if (statusCode == HttpStatus.SC_NOT_FOUND) {
-                return false;
-            } else {
-                String defaultMessage = String.format("Checking entry from '%s' response status %d: %s", safeUri(uri), statusCode, statusLine.getReasonPhrase());
-                return throwHttpStatusCodeException(statusCode, defaultMessage);
-            }
-        } catch (IOException e) {
-            throw wrap(e);
-        }
     }
 
     @Override
@@ -128,7 +103,7 @@ public class HttpBuildCacheService implements NextGenBuildCacheService {
     }
 
     @Override
-    public void store(BuildCacheKey key, NextGenWriter writer) throws BuildCacheException {
+    public void store(BuildCacheKey key, BuildCacheEntryWriter writer) throws BuildCacheException {
         final URI uri = root.resolve(key.getHashCode());
         HttpPut httpPut = new HttpPut(uri);
         if (useExpectContinue) {
