--- subprojects/build-cache/src/main/java/org/gradle/caching/internal/NoOpBuildCacheService.java
@@ -1,54 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.caching.internal;
-
-import org.gradle.caching.BuildCacheEntryReader;
-import org.gradle.caching.BuildCacheEntryWriter;
-import org.gradle.caching.BuildCacheException;
-import org.gradle.caching.BuildCacheKey;
-import org.gradle.caching.BuildCacheService;
-
-import java.io.IOException;
-
-public class NoOpBuildCacheService implements NextGenBuildCacheService {
-    public static final BuildCacheService INSTANCE = new NoOpBuildCacheService();
-
-    private NoOpBuildCacheService() {
-    }
-
-    @Override
-    public boolean contains(BuildCacheKey key) {
-        return false;
-    }
-
-    @Override
-    public boolean load(BuildCacheKey key, BuildCacheEntryReader reader) throws BuildCacheException {
-        return false;
-    }
-
-    @Override
-    public void store(BuildCacheKey key, BuildCacheEntryWriter legacyWriter) throws BuildCacheException {
-    }
-
-    @Override
-    public void store(BuildCacheKey key, NextGenWriter writer) throws BuildCacheException {
-    }
-
-    @Override
-    public void close() throws IOException {
-    }
-}
